Upgrade to LLVM 7.0

+++ test/test087.c
@@ -17,9 +17,9 @@ int main() {
   char inc0 = 0, inc1 = 0;
   int diff = 0, a = 100;
   int *p = &a;
-  inc0 =(int)p;
+  inc0 =(long)p;
   --(*p--);  //--*p--);
-  inc1 =(int)p;
+  inc1 =(long)p;
   diff =  inc0-inc1;
   diff += 2;
   return diff;
